I added the prediction dictionary in addition to the weight data.
For random forest and logistic regression, the contents of the dictionaries include column names, scores, and predictions.

For GaussianNB, there is no weight information, and the prediction dictionary includes only scores and predictions.

For other models, there is weight information, and the contents of the prediction dictionaries include scores and predictions only.